Moved style to better place - not a component
