+++ src/components/emergency.js
@@ -3,7 +3,7 @@ import { Dialog } from '@reach/dialog'
 import VisuallyHidden from 'components/visually-hidden'
 import styled from '@emotion/styled'
 import '@reach/dialog/styles.css'
-import { colors } from 'components/styles/theme'
+import { colors } from 'style/theme'
 import color from 'color'
 
 const EmergencyDialog = styled(Dialog)`
