+++ src/components/footer/site.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from '@emotion/styled'
-import { colors } from 'components/styles/theme'
+import { colors } from 'style/theme'
 import { Flex, Box } from '@rebass/grid/emotion'
 import Container from 'components/container'
 import Link from 'gatsby-link'
