+++ src/components/header/applicant.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from '@emotion/styled'
-import { colors } from 'components/styles/theme'
+import { colors } from 'style/theme'
 import { UserContext } from 'components/contexts/user'
 import Link from 'gatsby-link'
