+++ src/components/header/page-title.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from '@emotion/styled'
-import { colors } from 'components/styles/theme'
+import { colors } from 'style/theme'
 
 const PageTitleHeader = styled('h1')`
   color: ${colors.primary.darkest};
