+++ src/components/header/site-header.js
@@ -2,8 +2,8 @@ import React from 'react'
 import styled from '@emotion/styled'
 import Link from 'gatsby-link'
 import Container from 'components/container'
-import { colors } from 'components/styles/theme'
-import bp from 'components/styles/breakpoints'
+import { colors } from 'style/theme'
+import bp from 'style/breakpoints'
 
 const siteHeaderSize = bp({
   fontSize: ['1.5rem', '2rem', '2.5rem'],
