+++ src/components/homepages/samples/lazy-hero.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import LazyHero from 'react-lazy-hero'
-import { colors, fonts } from 'components/styles/theme'
+import { colors } from 'style/theme'
 import styled from '@emotion/styled'
 import Link from 'gatsby-link'
 import { LeadParagraph } from 'components/type'
