+++ src/components/pages/event.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { Flex, Box } from '@rebass/grid/emotion'
 import Container from 'components/container'
 import { LeadParagraph } from 'components/type'
-import { colors } from 'components/styles/theme'
+import { colors } from 'style/theme'
 import Link from 'gatsby-link'
 import styled from 'react-emotion'
 import moment from 'moment'
