+++ src/pages/academics/index.js
@@ -4,7 +4,7 @@ import Link from 'gatsby-link'
 import { graphql } from 'gatsby'
 import { InputText, Submit } from 'components/forms'
 import styled from '@emotion/styled'
-import { colors } from 'components/styles/theme'
+import { colors } from 'styles/theme'
 import { Flex, Box } from '@rebass/grid/emotion'
 import { LeadParagraph } from 'components/type'
 import { UnstyledList } from 'components/type'
