+++ src/stories/colors.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import typography from 'utils/typography'
-import { colors } from 'components/styles/theme'
+import { colors } from 'styles/theme'
 import styled from '@emotion/styled'
 import { Flex, Box } from '@rebass/grid/emotion'
